refactor(web): extract browser DOM backend into rinch-web crate
The browser-native DOM backend (WebDocument over web_sys) plus its event-
delegation and mount glue lived only in examples/ui-zoo-web and had been
forked into examples/paint-web, where it silently drifted: paint-web's copy
was missing the 7 scroll/client DomDocument methods and wired only a `click`
listener (no mousedown/mousemove/mouseup), leaving the embedded Slider/
ColorPicker drag non-functional.

Reclaim the `rinch-web` name (previously a dead, unreferenced vello/canvas
stub) for the real backend:

- Delete the stub (event_loop/renderer/window) and its dead facade wiring
  (rinch's optional `rinch-web` dep + `web` feature), which also removes the
  would-be rinch -> rinch-web -> rinch cycle.
- crates/rinch-web now holds the ui-zoo-web superset: WebDocument,
  setup_event_delegation (mousedown/move/up + keydown + input, text-hit,
  data-onsubmit), and a composable `mount(theme, build)` helper.
- Migrate ui-zoo-web and paint-web onto it; drop their now-transitive
  js-sys/web-sys deps. paint-web gains the scroll API + working drag.
- Update README/wasm guide/AGENTS/overview docs to the rinch_web::mount API.

rinch-web is an in-workspace member, so the web DOM path gets native
cargo check/clippy coverage for the first time.

Verified: cargo build --target wasm32-unknown-unknown for both examples,
cargo check --workspace, cargo clippy, rustfmt --check.
