+++ crates/rinch-web/Cargo.toml
@@ -3,19 +3,25 @@ name = "rinch-web"
 version.workspace = true
 edition.workspace = true
 license.workspace = true
-description = "Web backend for rinch (WASM target)"
+description = "Browser-native DOM backend for rinch (WASM target): WebDocument, event delegation, and mount"
 
 [dependencies]
-rinch-platform = { workspace = true }
-vello = { workspace = true }
-peniko = "0.6"
-web-sys = { version = "0.3", features = [
-    "Window", "Document", "HtmlCanvasElement", "Element",
-    "MouseEvent", "KeyboardEvent", "WheelEvent",
-    "EventTarget", "AddEventListenerOptions",
-    "Performance", "PerformanceTiming",
-    "console",
-] }
+# The backend implements rinch-core's DomDocument trait; the event-delegation
+# and mount glue use the rinch facade (render_surface, setup_theme_css) plus
+# rinch-core's events/drag/render-scope. No rinch-dom/Taffy/Parley/Vello — the
+# browser handles layout, text, CSS, and painting natively.
+rinch-core = { workspace = true, features = ["theme"] }
+rinch = { workspace = true, default-features = false, features = ["components", "theme"] }
+
+# WASM bindings
 wasm-bindgen = "0.2"
 js-sys = "0.3"
-tracing = { workspace = true }
+web-sys = { version = "0.3", features = [
+    "Window", "Document", "Element", "HtmlElement", "HtmlInputElement",
+    "HtmlTextAreaElement", "HtmlHeadElement",
+    "Node", "NodeList", "Text", "Comment",
+    "CssStyleDeclaration", "DomRect",
+    "Event", "MouseEvent", "KeyboardEvent",
+    "EventTarget",
+    "Range",
+] }
