+++ crates/rinch/Cargo.toml
@@ -46,7 +46,6 @@ rinch-theme = { workspace = true, optional = true }
 rinch-components = { workspace = true, optional = true }
 rinch-debug = { path = "../rinch-debug", optional = true }
 rinch-editor = { workspace = true }
-rinch-web = { workspace = true, optional = true }
 rinch-video = { workspace = true, optional = true }
 ureq = { version = "3", optional = true }
 memory-stats = { version = "1.2", optional = true }
@@ -114,7 +113,6 @@ memory-profile = ["memory-stats"]
 theme = ["dep:rinch-theme", "rinch-core/theme"]
 components = ["theme", "dep:rinch-components"]
 debug = ["dep:rinch-debug"]
-web = ["dep:rinch-web"]
 video = ["dep:rinch-video"]
 image-network = ["dep:ureq"]
 collaboration = ["rinch-editor/collaboration"]
